+++ CHANGES.md
@@ -1,3 +1,10 @@
+V3.0.0
+- mappMotion 5.26.x required as minimum
+- TD_WebServices: Remove output SegInfo.segmentCyclicInfo
+- TD_WebServices: Add output SegInfo.segmentData
+- TD_WebServices: Segments are colored red when in ErrorStop
+
+
 V2.7.2
 - TD_Recorder & TD_Webservices: the shuttles are drawn last so that they are not overdrawn.